+++ packages/fixie-common/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @fixieai/fixie-common
 
+## 1.0.6
+
+### Patch Changes
+
+- Log voice worker latency.
+
+## 1.0.5
+
+### Patch Changes
+
+- Fix problem with FixieAgentBase.update using wrong field name.
+
+## 1.0.4
+
+### Patch Changes
+
+- Fix createRevision.
+
+## 1.0.3
+
+### Patch Changes
+
+- Add additional deployment parameters to Fixie types.
+
 ## 1.0.2
 
 ### Patch Changes